Final commit (see full message)
There are still a few issues with the application. React Router was not my friend today: If you refresh any routes that the interface provides, the content renders properly, but if you're trying to view things from the list view between paginated colors and colors by name, the UI won't sync with the router state. I also didn't implement a search function nor a quick way to navigate to the landing page. Lastly, if you click on a page number on the 10s, the nav list fails to
render.